// src/app/features/maint-rate-summary-query/models/summary-detail.model.ts
export interface SummaryDetail {
  ttype: string;
  ttypename: string;
  stock: string;
  stocknm: string;
  bqty: string;
  price: string;
  lprice: string;
  bcramt: string;
  bdnamt: string;
  bgtamt: string;
  agpamt: string;
  collateral: string;
  keeprate: string;
  sftype: string;
  tdate: string;
  dseq: string;
}